Nation entity refactored to have Institution inside it

+++ worldcup/src/main/java/com/medinastr/worldcup/service/NationService.java
@@ -5,6 +5,7 @@ import com.medinastr.worldcup.dto.NationDTO;
 import com.medinastr.worldcup.entity.Nation;
 import com.medinastr.worldcup.exception.WorldcupConflictException;
 import com.medinastr.worldcup.exception.WorldcupInvalidAttributeException;
+import com.medinastr.worldcup.exception.WorldcupNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -30,32 +31,40 @@ public class NationService {
     }
 
     public Nation saveNation(NationDTO nationDTO) {
-        validateNationName(nationDTO.getName());
-        Nation dbNation = nationDTO.toNation();
-        return nationRepository.save(dbNation);
+        validateNation(nationDTO);
+        return nationRepository.save(nationDTO.toNation());
     }
 
     public List<Nation> saveNationsList(List<NationDTO> nationsDTO) {
         List<Nation> dbNations = nationsDTO.stream()
                 .map(nationDTO -> {
-                    validateNationName(nationDTO.getName());
+                    validateNation(nationDTO);
                     return nationDTO.toNation();
                 })
                 .collect(Collectors.toList());
         return nationRepository.saveAll(dbNations);
     }
 
-    public void validateNationName(String name) {
-        if (name == null || name.length() < 3 || !name.matches("[a-zA-Z\\s]+")) {
-            throw new WorldcupInvalidAttributeException("Nation name invalid.");
-        } else if (nationRepository.findByName(name).isPresent()) {
+    public void validateNation(NationDTO nationDTO) {
+        if (nationDTO.getName() == null || nationDTO.getName().length() < 3
+                || !nationDTO.getName().matches("[a-zA-Z\\s]+")) {
+            throw new WorldcupInvalidAttributeException("Name invalid.");
+        }
+        if (nationRepository.findByName(nationDTO.getName()).isPresent()) {
             throw new WorldcupConflictException("Nation already exists.");
+        }
+        if (nationDTO.getInstitution() == null || nationDTO.getInstitution().length() < 2
+                || !nationDTO.getInstitution().matches("[a-zA-Z\\s]+")) {
+            throw new WorldcupInvalidAttributeException("Institution name invalid.");
         }
     }
 
     public Optional<Nation> delete(int id) {
-        Optional<Nation> findNation = nationRepository.findById(id);
+        Optional<Nation> dbNation = nationRepository.findById(id);
+        if(dbNation.isEmpty()) {
+            throw new WorldcupNotFoundException("Nation not exists.");
+        }
         nationRepository.deleteById(id);
-        return findNation;
+        return dbNation;
     }
 }
